Remove sending status after every single step
In previous versions of AVaRICE, a complete processor status was sent
after every instruction, apparently to prevent GDB from asking for the
status itself. In GDB 7.6, this is not required: GDB only asks for the
status after a code line (not instruction) was finished. By removing the
frequent status read in AVaRICE, debugging can be sped up.