Implement setBreakOnChangeOfFlow
To properly implement range stepping later on, the debugger must be
notified of unexpected branches during execution. This parameter may
only be set on mkII devices, as those are the only ones that possibly
can support range stepping.

Leaving the "break on change of flow" flag enabled does not influence
other debug commands like continue: The JTAG ICE does report a break to
GDB, but GDB itself continues, as long as the breakpoint has not been
reached.